+++ src/main/java/com/keepitup/magjobbackend/rolemember/controller/impl/RoleMemberDefaultController.java
@@ -1,7 +1,10 @@
 package com.keepitup.magjobbackend.rolemember.controller.impl;
 
+import com.keepitup.magjobbackend.configuration.Constants;
+import com.keepitup.magjobbackend.configuration.SecurityService;
 import com.keepitup.magjobbackend.member.entity.Member;
 import com.keepitup.magjobbackend.member.service.impl.MemberDefaultService;
+import com.keepitup.magjobbackend.organization.entity.Organization;
 import com.keepitup.magjobbackend.role.entity.Role;
 import com.keepitup.magjobbackend.role.service.impl.RoleDefaultService;
 import com.keepitup.magjobbackend.rolemember.controller.api.RoleMemberController;
@@ -28,6 +31,7 @@ public class RoleMemberDefaultController implements RoleMemberController {
     private final RoleMemberToResponseFunction roleMemberToResponseFunction;
     private final RoleMembersToResponseFunction roleMembersToResponseFunction;
     private final RequestToRoleMemberFunction requestToRoleMemberFunction;
+    private final SecurityService securityService;
 
     @Autowired
     public RoleMemberDefaultController(
@@ -36,7 +40,8 @@ public class RoleMemberDefaultController implements RoleMemberController {
             MemberDefaultService memberService,
             RoleMemberToResponseFunction roleMemberToResponseFunction,
             RoleMembersToResponseFunction roleMembersToResponseFunction,
-            RequestToRoleMemberFunction requestToRoleMemberFunction
+            RequestToRoleMemberFunction requestToRoleMemberFunction,
+            SecurityService securityService
     ) {
         this.roleMemberService = roleMemberService;
         this.roleService = roleService;
@@ -44,15 +49,23 @@ public class RoleMemberDefaultController implements RoleMemberController {
         this.roleMemberToResponseFunction = roleMemberToResponseFunction;
         this.roleMembersToResponseFunction = roleMembersToResponseFunction;
         this.requestToRoleMemberFunction = requestToRoleMemberFunction;
+        this.securityService = securityService;
     }
 
     @Override
     public GetRoleMembersResponse getRoleMembers() {
+        if (!securityService.hasAdminPermission()) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         return roleMembersToResponseFunction.apply(roleMemberService.findAll());
     }
 
     @Override
     public GetRoleMemberResponse getRoleMember(BigInteger id) {
+        if (!securityService.hasAdminPermission()) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         return roleMemberService.find(id)
                 .map(roleMemberToResponseFunction)
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
@@ -65,6 +78,12 @@ public class RoleMemberDefaultController implements RoleMemberController {
         Role role = roleOptional
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
 
+        Organization organization = role.getOrganization();
+
+        if(!securityService.belongsToOrganization(organization)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         return roleMembersToResponseFunction.apply(roleMemberService.findAllByRole(role));
     }
 
@@ -75,11 +94,25 @@ public class RoleMemberDefaultController implements RoleMemberController {
         Member member = memberOptional
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
 
+        Organization organization = member.getOrganization();
+
+        if(!securityService.belongsToOrganization(organization)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         return roleMembersToResponseFunction.apply(roleMemberService.findAllByMember(member));
     }
 
     @Override
     public GetRoleMemberResponse createRoleMember(PostRoleMemberRequest postRoleMemberRequest) {
+        Organization organization = roleService.find(postRoleMemberRequest.getRole()).orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
+        ).getOrganization();
+
+        if (!securityService.hasPermission(organization, Constants.PERMISSION_NAME_CAN_MANAGE_ROLES)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         roleMemberService.create(requestToRoleMemberFunction.apply(postRoleMemberRequest));
 
         return roleMemberService.findByMemberAndRole(
@@ -95,6 +128,16 @@ public class RoleMemberDefaultController implements RoleMemberController {
 
     @Override
     public void deleteRoleMember(BigInteger id) {
+        Organization organization = roleService.find(roleMemberService.find(id).orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
+        ).getRole().getId()).orElseThrow(
+                () -> new ResponseStatusException(HttpStatus.NOT_FOUND)
+        ).getOrganization();
+
+        if (!securityService.hasPermission(organization, Constants.PERMISSION_NAME_CAN_MANAGE_ROLES)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         roleMemberService.find(id)
                 .ifPresentOrElse(
                         roleMember -> roleMemberService.delete(id),
